# back-end/swagger.yaml
openapi: 3.0.4
info:
  title: Healthcare Management API
  description: |
    API documentation for Healthcare Management System
    
    This API provides endpoints for user authentication, user management, and healthcare-related operations.
    
    ## Authentication
    Most endpoints require authentication using JWT tokens. Include the token in the Authorization header:
   
    ## User Roles
    - `admin`: Full system access
    - `manager`: Management access
    - `user`: Basic user access
    - `service`: Service provider access
    - `consultant`: Consultant access
    
    ## User Status
    - `0`: Inactive (soft deleted)
    - `1`: Active
    - `2`: Locked
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
  /roles:
    get:
      tags:
        - Roles
      summary: List roles
      description: Get a paginated list of roles
      operationId: listRoles
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [name, code, createdAt]
        - name: sortOrder
          in: query
          required: false
          schema:
            type: integer
            enum: [1, -1]
            example: 1
        - name: page
          in: query
          required: false
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Roles retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Get roles successfully"
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Role'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer

    post:
      tags:
        - Roles
      summary: Create role
      description: Create a new role
      operationId: createRole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - code
              properties:
                name:
                  type: string
                  example: "Laboratory Manager"
                code:
                  type: string
                  example: "manager"
                description:
                  type: string
                privileges:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Role created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Role created successfully"
                  data:
                    $ref: '#/components/schemas/Role'
        '409':
          description: Role code already exists
        '422':
          description: Validation error
  /roles/{id}:
    put:
      tags:
        - Roles
      summary: Update role
      description: Update a role by id
      operationId: updateRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                privileges:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Role updated successfully"
                  data:
                    $ref: '#/components/schemas/Role'
        '404':
          description: Role not found
        '422':
          description: Validation error


    delete:
      tags:
        - Roles
      summary: Delete role
      description: Delete a role from role collection
      operationId: deleteRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Role ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        '200':
          description: Role deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Role deleted successfully"
                  data:
                    $ref: '#/components/schemas/Role'
        '400':
          description: Invalid role id
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Role not found
        '422':
          description: Validation error
  # Authentication endpoints
...
  /event-logs:
    get:
      tags:
        - Event Logs
      summary: List event logs
      description: Retrieve a paginated list of event logs with optional filters
      operationId: listEventLogs
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: action
          in: query
          required: false
          schema:
            type: string
        - name: operatorName
          in: query
          required: false
          schema:
            type: string
        - name: operatorRole
          in: query
          required: false
          schema:
            type: string
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [timestamp, action]
        - name: sortOrder
          in: query
          required: false
          schema:
            type: integer
            enum: [1, -1]
            example: -1
        - name: page
          in: query
          required: false
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            example: 20
      responses:
        '200':
          description: Event logs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Event logs retrieved"
                  data:
                    type: object
                    properties:
                      eventLogs:
                        type: array
                        items:
                          $ref: '#/components/schemas/EventLog'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          limit:
                            type: integer
                          total:
                            type: integer
                          totalPages:
                            type: integer
  /event-logs/{id}:
    get:
      tags:
        - Event Logs
      summary: Get event log by id
      description: Retrieve a single event log by its id
      operationId: getEventLogById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        '200':
          description: Event log retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Event log retrieved"
                  data:
                    $ref: '#/components/schemas/EventLog'
        '400':
          description: Invalid id
        '401':
          description: Unauthorized
        '404':
          description: Event log not found
        '422':
          description: Validation error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    Role:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439022"
        name:
          type: string
          example: "Administrator"
        code:
          type: string
          example: "admin"
        description:
          type: string
          example: "Access all features"
        privileges:
          type: array
          items:
            type: string
          example: ["read_only"]
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00.000Z"
...
    Operator:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "Nguyễn Văn A"
        role:
          type: string
          example: "manager"
    EventLog:
      type: object
      properties:
        _id:
          type: string
          example: "60d5f9b2b60a3c6f1c9e8b7a"
        operator:
          $ref: '#/components/schemas/Operator'
        action:
          type: string
          example: "USER_UPDATED"
        details:
          type: string
          example: "User information updated"
        timestamp:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00.000Z"